Storybook: Fix library resolution; store and global variables declaration
Storybook: Clean up store import

Storybook: Convert store global variable syntax to Vue3 savvy
