+++ storybook/.storybook/preview.js
@@ -1,21 +1,13 @@
 import  { setup }  from '@storybook/vue3';
-import { createStore } from 'vuex'
 import vSelect from 'vue-select';
 import FloatingVue from 'floating-vue';
 import { themes } from '@storybook/theming';
-import { get } from '../../shell/utils/object';
-import IntlMessageFormat from 'intl-messageformat';
 import installShortcut from './theme-shortcut';
 import withEvents from 'storybook-auto-events';
-const i18nStrings = require('../../shell/assets/translations/en-us.yaml');
 import cleanTooltipDirective  from '@shell/directives/clean-tooltip';
 import ShortKey from 'vue-shortkey';
 import cleanHtmlDirective from '@shell/directives/clean-html';
-
-// Store modules
-import growl from './store/growl';
-import codeMirror from './store/codeMirror';
-import table from './store/table';
+import store from './store'
 
 // i18n
 import i18n from '../../shell/plugins/i18n';
@@ -32,36 +24,16 @@ setup((vueApp) => {
     template: '<a>link</a>',
   })
 
-  vueApp.use(storePlugin);
+  vueApp.use({
+    store,
+    install(vueApp) {
+      // Vuex is used through the app with a global variable
+      vueApp.config.globalProperties.$store = store
+    },
+  });
 
   window.__codeMirrorLoader = () => import(/* webpackChunkName: "codemirror" */ '@shell/plugins/codemirror');
 })
-
-const storePlugin = {
-  store: createStore({
-    getters: {
-      'i18n/exists': key => store.getters['i18n/t'],
-      'i18n/t': state => (key, args) => {
-        const msg = get(i18nStrings, key) || key;
-  
-        if (msg?.includes('{')) {
-          const formatter = new IntlMessageFormat(msg, state.selected);
-          return formatter.format(args);
-        }
-  
-        return msg;
-      },
-    },
-    modules: {
-      growl,
-      codeMirror,
-      table
-    }
-  }),
-  install (vueApp, options) {
-    vueApp.prototype.$store = store
-  }
-};
 
 const preview = {
   parameters: {
